requirements.txt
apache-beam==2.43.0
apache-flink==1.17.0
apache-flink-libraries==1.17.0
avro-python3==1.9.2.1
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.0
charset-normalizer==3.1.0
cloudpickle==2.2.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
Faker==22.2.0
fastavro==1.4.7
fasteners==0.18
find-libpython==0.3.0
grpcio==1.54.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
jieba @ file:///home/conda/feedstock_root/build_artifacts/jieba_1622403007736/work
kafka-python @ file:///home/conda/feedstock_root/build_artifacts/kafka-python_1601470752419/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy==1.21.6
objsize==0.5.2
orjson==3.8.11
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.3.5
pemja==0.3.0
Pillow==9.1.0
proto-plus==1.22.2
protobuf==3.20.3
py4j==0.10.9.7
pyarrow==8.0.0
pycparser==2.21
pydot==1.4.2
pyheif==0.7.0
pymongo==3.13.0
pyparsing==3.0.9
pyspark @ file:///croot/pyspark_1670256141824/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
regex==2023.5.5
requests==2.30.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
typing_extensions==4.5.0
urllib3==2.0.2
whatimage==0.0.3
zstandard==0.21.0
